+++ crm1/src/main/java/com/sc/service/XtUserAccountService.java
@@ -1,7 +1,5 @@
 package com.sc.service;
 
-import java.math.BigDecimal;
-
 import com.github.pagehelper.PageInfo;
 import com.sc.entity.XtUserAccount;
 
@@ -11,9 +9,9 @@ public interface XtUserAccountService
 	
 	public void updateXtUserAccount(XtUserAccount UserAccount);
 	
-	public void deleteXtUserAccount(BigDecimal id);
+	public void deleteXtUserAccount(Long id);
 	
-	public XtUserAccount getXtUserAccount(BigDecimal userId);
+	public XtUserAccount getXtUserAccount(Long userId);
 	
 	public PageInfo<XtUserAccount> seleteXtUserAccount(Integer pageNum,Integer pageSize,XtUserAccount UserAccount);
